refactor: fetchCurrencyData module
Upon completion of the AWS Lambda function fetchCurrencyData, it will trigger another AWS Lambda function within a Step Function workflow.
This change will allow the second AWS Lambda function to trigger AWS Glue jobs to process and refine the ingested data.
The LoadService has a new property that stores a list of uploaded objects to be returned as data by the AWS Lambda run.
The fetchCurrencyData AWS Function now returns a payload with the list of uploaded objects