docs: Create notebooks coincap_api_processing_data and coincap_api
The notebook coincap_api_processing_data tests the treating and refining of data fetched from API CoinCap using Apache Spark (PySpark) for the job.
The notebook coincap_api was used to visualize the different API CoinCap's endpoints and to explore the data.
Both of the notebooks were used to prove some concepts in the early development of the project.