fix: fetchCurrencyData module
Fixed the filename set in load.py.
Modified the execute function in etl.py to return a string with a list of object keys in the uploaded_object instead of a list
This change was necessary for integration with a Step Functions workflow where the list of object keys must be a string to be passed as a parameter.
Modified the lambda function to return a JSON object instead of a stringified JSON object. The stringified JSON object was causing an error in the workflow.